+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 # 1
 
-## 1.6
+## 1.6.17
 
  * Add preloads configuration option to lua_web/lua_general modules
  * Add mtev_set_app_name and mtev_get_app_name for conf file root discovery.